+++ etna/metrics/functional_metrics.py
@@ -195,6 +195,8 @@ def smape(y_true: ArrayLike, y_pred: ArrayLike, eps: float = 1e-15, multioutput:
     .. math::
        SMAPE(y\_true, y\_pred) = \\frac{2 \\cdot 100 \\%}{n} \\cdot \\sum_{i=1}^{n} \\frac{\\mid y\_true_i - y\_pred_i\\mid}{\\mid y\_true_i \\mid + \\mid y\_pred_i \\mid}
 
+    The nans are ignored during computation. If all values are nans, the result is NaN.
+
     Parameters
     ----------
     y_true:
@@ -247,6 +249,8 @@ def sign(y_true: ArrayLike, y_pred: ArrayLike, multioutput: str = "joint") -> Ar
     .. math::
         Sign(y\_true, y\_pred) = \\frac{1}{n}\\cdot\\sum_{i=1}^{n}{sign(y\_true_i - y\_pred_i)}
 
+    The nans are ignored during computation. If all values are nans, the result is NaN.
+
     Parameters
     ----------
     y_true:
@@ -275,8 +279,15 @@ def sign(y_true: ArrayLike, y_pred: ArrayLike, multioutput: str = "joint") -> Ar
         raise ValueError("Shapes of the labels must be the same")
 
     axis = _get_axis_by_multioutput(multioutput)
+    with warnings.catch_warnings():
+        # this helps to prevent warning in case of all nans
+        warnings.filterwarnings(
+            message="Mean of empty slice",
+            action="ignore",
+        )
+        result = np.nanmean(np.sign(y_true_array - y_pred_array), axis=axis)
 
-    return np.mean(np.sign(y_true_array - y_pred_array), axis=axis)
+    return result
 
 
 def max_deviation(y_true: ArrayLike, y_pred: ArrayLike, multioutput: str = "joint") -> ArrayLike:
@@ -284,6 +295,8 @@ def max_deviation(y_true: ArrayLike, y_pred: ArrayLike, multioutput: str = "join
 
     .. math::
         MaxDeviation(y\_true, y\_pred) = \\max_{1 \\le j \\le n} | y_j |, where \\, y_j = \\sum_{i=1}^{j}{y\_pred_i - y\_true_i}
+
+    The nans are ignored during computation. If all values are nans, the result is NaN.
 
     Parameters
     ----------
@@ -313,9 +326,15 @@ def max_deviation(y_true: ArrayLike, y_pred: ArrayLike, multioutput: str = "join
         raise ValueError("Shapes of the labels must be the same")
 
     axis = _get_axis_by_multioutput(multioutput)
-
-    prefix_error_sum = np.cumsum(y_pred_array - y_true_array, axis=axis)
-    return np.max(np.abs(prefix_error_sum), axis=axis)
+    diff = y_pred_array - y_true_array
+    prefix_error_sum = np.nancumsum(diff, axis=axis)
+    isnan = np.all(np.isnan(diff), axis=axis)
+    result = np.max(np.abs(prefix_error_sum), axis=axis)
+    result = np.where(isnan, np.NaN, result)
+    try:
+        return result.item()
+    except ValueError as e:
+        return result  # type: ignore
 
 
 rmse = partial(mse_sklearn, squared=False)
@@ -327,6 +346,8 @@ def wape(y_true: ArrayLike, y_pred: ArrayLike, multioutput: str = "joint") -> Ar
     .. math::
         WAPE(y\_true, y\_pred) = \\frac{\\sum_{i=1}^{n} |y\_true_i - y\_pred_i|}{\\sum_{i=1}^{n}|y\\_true_i|}
 
+    The nans are ignored during computation. If all values are nans, the result is NaN.
+
     Parameters
     ----------
     y_true:
@@ -355,8 +376,35 @@ def wape(y_true: ArrayLike, y_pred: ArrayLike, multioutput: str = "joint") -> Ar
         raise ValueError("Shapes of the labels must be the same")
 
     axis = _get_axis_by_multioutput(multioutput)
-
-    return np.sum(np.abs(y_true_array - y_pred_array), axis=axis) / np.sum(np.abs(y_true_array), axis=axis)  # type: ignore
+    diff = y_true_array - y_pred_array
+    numerator = np.nansum(np.abs(diff), axis=axis)
+    isnan = np.isnan(diff)
+    denominator = np.nansum(np.abs(y_true_array * (~isnan)), axis=axis)
+    with warnings.catch_warnings():
+        # this helps to prevent warning in case of all nans
+        warnings.filterwarnings(
+            message="invalid value encountered in scalar divide",
+            action="ignore",
+        )
+        warnings.filterwarnings(
+            message="invalid value encountered in divide",
+            action="ignore",
+        )
+        warnings.filterwarnings(
+            message="divide by zero encountered in scalar divide",
+            action="ignore",
+        )
+        warnings.filterwarnings(
+            message="divide by zero encountered in divide",
+            action="ignore",
+        )
+        isnan = np.all(isnan, axis=axis)
+        result = np.where(denominator == 0, np.NaN, numerator / denominator)
+        result = np.where(isnan, np.NaN, result)
+        try:
+            return result.item()
+        except ValueError as e:
+            return result  # type: ignore
 
 
 def count_missing_values(y_true: ArrayLike, y_pred: ArrayLike, multioutput: str = "joint") -> ArrayLike: